+++ src/lib/ngx-jsonld-provider.component.ts
@@ -1,0 +1,19 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'ngx-ngx-jsonld-provider',
+  template: `
+    <p>
+      ngx-jsonld-provider works!
+    </p>
+  `,
+  styles: []
+})
+export class NgxJsonldProviderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}