+++ src/lib/ngx-jsonld-provider.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { NgxJsonldProviderComponent } from './ngx-jsonld-provider.component';
+
+@NgModule({
+  imports: [
+  ],
+  declarations: [NgxJsonldProviderComponent],
+  exports: [NgxJsonldProviderComponent]
+})
+export class NgxJsonldProviderModule { }