+++ src/public_api.ts
@@ -2,9 +2,9 @@
  * Public API Surface of ngx-jsonld
  */
 
-export { NgxJsonLd } from './lib/ngx-jsonld.service';
+export { NgxJsonLdProvider } from './lib/ngx-jsonld-provider.service';
+export { JsonLdDocument } from './lib/jsonld-document-model';
 export { CompactOptions, DocumentLoader, ExpandOptions, FlattenOptions,
          FrameOptions, FromRdfOptions, JsonLd, JsonLdObject, NormalizeOptions,
-         RDFParser, ToRdfOptions } from 'jsonld';
-// export * from './lib/ngx-jsonld.component';
-// export * from './lib/ngx-jsonld.module';
+         RDFParser, ToRdfOptions, JsonLdPrimitive } from 'jsonld';
+