+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ngx-jsonld Changelog
 
+## 1.0.0 (25 Sept 2018)
+
+Added instructions about using polyfill to run jsonld.js on Angular 6+.
+
+### 0.2.0 (24 Sept 2018)
+
+Added missing imports.
+
 ### 0.1.0 (24 Sept 2018)
 
 First commit.